refactor(src/bot src/sources): sending new BotError when bot or stream source has failed
